Fix register and shift amount assembling.
Also changes :if-exists :overwrite to :if-exists :supersede as the
former simply overwrites the file in place, and can leave behind traces
of the previous output if it extended beyond the current output (such as
assembling a smaller assembly source file with the same name as a longer
assembly source file).
Supersede creates a new file entirely, and replaces the old one.